move spring show to separate file
